增加房型庫存repo & entity，並且修改價格型別

+++ src/main/java/com/example/petel/dto/BOOKTranrqOrderDetail.java
@@ -34,6 +34,6 @@ public class BOOKTranrqOrderDetail implements Serializable {
 
     @JsonProperty("product_price")
     @NotNull(message = "product_price不得為空")
-    private Double productPrice;
+    private Integer productPrice;
 }
